requirements.txt
requests
python-dotenv
scikit-learn
spacy
openai
langchain
langchain-openai
langchain-community
langchain-huggingface